Refactored CheckController. Split it in two so as two have two distinct classes one UserController and one CheckController
